changed video div to popUp
