+++ src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import Hero from '../components/Hero/index.astro';
+import Hero from '../components/HomePage/Hero/index.astro';
 ---
 
 <Layout title="Welcome to TCET Mumbai" description='Thakur College of Engineeringa and Technology' >
